lab12

+++ app/Http/Controllers/ShopController.php
@@ -7,6 +7,7 @@ use App\Models\Shop;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 use Psr\Http\Message\ServerRequestInterface;
 
@@ -47,11 +48,13 @@ class ShopController extends SearchableController
     }
 
     function CreateForm(): View {
+    Gate::authorize('create',Shop::class);
     return view('shops.create-form');
     }
 
     function create(ServerRequestInterface $request): RedirectResponse {
     $shop = Shop::create($request->getParsedBody());
+    Gate::authorize('create',$shop);
 
     return redirect(
         session()->get('bookmarks.shops.create-form', route('shops.list'))
@@ -61,6 +64,7 @@ class ShopController extends SearchableController
 
     function UpdateForm(string $shopCode): View {
     $shop = $this->find($shopCode);
+    Gate::authorize('update',$shop);
 
     return view('shops.update-form', [
     'shop' => $shop,
@@ -72,6 +76,7 @@ class ShopController extends SearchableController
     string $shopCode,
     ): RedirectResponse {
     $shop = $this->find($shopCode);
+    Gate::authorize('update',$shop);
     $shop->fill($request->getParsedBody());
     $shop->save();
 
@@ -83,6 +88,7 @@ class ShopController extends SearchableController
 
     function delete(string $shopCode): RedirectResponse {
     $shop = $this->find($shopCode);
+    Gate::authorize('delete',$shop);
     $shop->delete();
 
     return redirect(
@@ -114,6 +120,7 @@ class ShopController extends SearchableController
         string $shopCode
     ): View {
         $shop = $this->find($shopCode);
+        Gate::authorize('create',$shop);
         $criteria = $productController->prepareCriteria($request->getQueryParams());
        $query = $productController
             ->getQuery()
@@ -138,6 +145,7 @@ class ShopController extends SearchableController
         string $shopCode
     ): RedirectResponse {
         $shop = $this->find($shopCode);
+        Gate::authorize('create',$shop);
         $data = $request->getParsedBody();
         $product = $productController
             ->getQuery()
@@ -160,6 +168,7 @@ class ShopController extends SearchableController
         string $shopCode,
     ): RedirectResponse {
         $shop = $this->find($shopCode);
+        Gate::authorize('create',$shop);
         $data = $request->getParsedBody();
        
         $product = $shop->products()
